check valid address for Bytes
