+++ libs/soroban-rlp/src/lib.rs
@@ -4,3 +4,4 @@ pub mod decoder;
 pub mod encoder;
 mod utils;
 pub mod messages;
+pub mod address_utils;
